third_party/blink/web_tests/external/wpt/svg/interact/scripted/defer-01-expected.txt

This is a testharness.js-based test.
[FAIL] defer script runs later
  assert_equals: script run status expected "deferred" but got "inline"
Harness: the test ran to completion.
